A five round game of rock-paper-scissors
index.js contains three essential functions, namely:

1. getComputerChoice() - Generates a random number from 0 - 2 then
choose a play for the computer based on the random number generated.
2. playRound() - This function takes two parameters, the computer's
choice and the player'choice. It compairs the arguments passed to it and
returns a string indicating who won the round.
3. game() - This function simulates five rounds of rock-paper-scissors.
It prompts the user for their input, get the computer's input, calls
playRound() and keep a record of the scores.