Added a graphical user interface for playing
The simple UI includes three buttons, each allows the player to select
an option (rock, paper or scissors). The buttons are in a container that
is wired with an event listener that determines the player's move then
plays one round of the game.

The winner is the first person to score 5 points. When the game is won,
the winner is announced and the play buttons are disabled. The restart
button will reset the scores and enable the buttons if they are
disabled.